requirements.txt
git+git://github.com/PrettyWood/pydantic.git@f/discriminated-union#egg=pydantic[email]
git+git://github.com/andrewbolster/sqlmodel@patch-1#egg=sqlmodel
pyyaml
sqlmodel
fastapi
aiofiles
python-socketio
uvicorn
figenv
sqlalchemy
databases[sqlite]
passlib[bcrypt]
python-jose
python-multipart
